+++ app/src/main/AndroidManifest.xml
@@ -32,6 +32,6 @@
       android:theme="@style/SplashScreenTheme">
     </activity>
     <activity android:name=".testing.BindableAdapterTestActivity"/>
-    <activity android:name=".activity.HtmlParserTestActivty"/>
+    <activity android:name=".testing.HtmlParserTestActivty"/>
   </application>
 </manifest>
